+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Show privacy statement on launch #454
+- Allow access to privacy statement from Settings / Preferences
 - Update list of available secure internet servers on startup #447
 - APIv3: Use prefer_tcp, fix accept header for '/connect'
 - Support for net_gateway / net_gateway_ipv6 in OpenVPN #440